Minor changes and updated README

+++ client/client_download.py
@@ -22,54 +22,51 @@ raft_nodes = config['raft_nodes']
 cache_dir = str(os.path.dirname(os.path.abspath(__file__))) + '/downloads/.cache/'
 
 def download():
-        global pool
-    # ch = input("Do you want to download something?(y/n) ")
-    # if ch == 'y':
-        # file_name = input("enter file name to download: ")
-        file_name = "team1_35KB_1543885532.srt"
+    global pool
+    ch = input("Do you want to download something?(y/n) ")
+    if ch == 'y':
+        file_name = input("enter file name to download: ")
         download_stub = file_transfer_pb2_grpc.DataTransferServiceStub(grpc.insecure_channel(choice(raft_nodes)))
 
         # get the file location
         try:
-            file_loc_info = download_stub.RequestFileInfo(file_transfer_pb2.FileInfo(fileName = file_name), timeout=0.1)
+            file_loc_info = download_stub.RequestFileInfo(file_transfer_pb2.FileInfo(fileName = file_name), timeout=0.2)
         except:
             for ip in raft_nodes:
                 try:
                     download_stub = file_transfer_pb2_grpc.DataTransferServiceStub(grpc.insecure_channel(ip))
-                    file_loc_info = download_stub.RequestFileInfo(file_transfer_pb2.FileInfo(fileName = file_name), timeout=0.1)
+                    file_loc_info = download_stub.RequestFileInfo(file_transfer_pb2.FileInfo(fileName = file_name), timeout=0.2)
                     break
                 except:
                     pass
         proxies = []
         for p in file_loc_info.lstProxy:
             proxies.append(p.ip + ":" + p.port)
-        # print(file_loc_info.fileName, file_loc_info.maxChunks, proxies, file_loc_info.isFileFound)
-        print(file_loc_info.isFileFound)
+        print(file_loc_info.fileName, file_loc_info.maxChunks, proxies, file_loc_info.isFileFound)
 
         # Download the file
-        # start = time()
+        start = time()
         if file_loc_info.isFileFound:
-            # print("Starting Download")
+            print("Starting Download")
             for chunk_id in range(file_loc_info.maxChunks):
                 pool.add_task(downloader, file_name, chunk_id, proxies)
             pool.wait_completion()
 
-            # print("Stitching the chunks together")
+            print("Stitching the chunks together")
             list_of_chunks = [f for f in os.listdir(cache_dir + file_name + '/') if isfile(join(cache_dir + file_name + '/',f))]
 
             for i in range(len(list_of_chunks)):
                 file_chunk = list_of_chunks[i].rsplit('_',1)
                 list_of_chunks[i] = (file_chunk[0], int(file_chunk[1]))
             list_of_chunks = sorted(list_of_chunks, key=lambda x: x[1])
-            # print(list_of_chunks)
             list_of_chunks = [s[0]+'_'+str(s[1]) for s in list_of_chunks]
             with open("downloads/" + file_name, "wb") as f:
                 for chunk in list_of_chunks:
                     with open(cache_dir + file_name + '/' + chunk, "rb") as ch:
                         f.write(ch.read())
-            # print("Done...")
-            # print(time()-start)
-            # print("Cleaning up...")
+            print("Done...")
+            print(time()-start)
+            print("Cleaning up...")
             shutil.rmtree(cache_dir + file_name + '/')
 
 
@@ -96,8 +93,8 @@ def downloader(file_name, chunk_id, proxies):
 
 
 def get_file_list():
-    # print("Getting the list of files...")
-    # start = time()
+    print("Getting the list of files...")
+    start = time()
     get_file_list_stub = file_transfer_pb2_grpc.DataTransferServiceStub(grpc.insecure_channel(choice(raft_nodes)))
     try:
         file_list = get_file_list_stub.ListFiles(file_transfer_pb2.RequestFileList(isClient = True))
@@ -109,16 +106,10 @@ def get_file_list():
                 break
             except:
                 pass
-    # print(time()-start)
+    print(time()-start)
     print(file_list.lstFileNames)
 
 
 if __name__ == "__main__":
-    s = time()
-    for i in range(1):
-        print(i)
-        get_file_list()
-        # download()
-    print(time()-s)
-    # get_file_list()
-    # download()
+    get_file_list()
+    download()
